tests/wpt/mozilla/tests/mozilla/bluetooth/requestDevice/two-filters.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(function() {
    window.testRunner.setBluetoothMockDataSet(adapter_type.glucose_heart_rate);
    return window.navigator.bluetooth.requestDevice(
        { filters: [{services: [battery_service.name]},
                    {services: [heart_rate.name]}]}
    ).then(glucose_device => {
        assert_equals(glucose_device.name, 'Heart Rate Device');
    })
    .catch(() => {
        assert_unreached('Promise should have resolved: An extra filter doesn\'t prevent matching.');
    })
}, 'An extra filter doesn\'t prevent matching.');
</script>
